[lldb] Distinguish a broken scripted provider from one that answers "no"
Several scripted-extension callbacks report absence with the same value
they'd report failure with. A synthetic child provider has no child of a
given name; one whose `get_child_index` raised also cannot produce an
index. Both arrived as "some error", so callers could not tell them apart
and had to guess - `ValueObjectSynthetic::GetIndexOfChildWithName` guessed
"absent" and fell back to automatic "[N]" subscripting, silently returning
a different child than the one asked for.

Type the failure instead, with `ScriptedExtensionError`, so a caller can
key off "the script is broken" rather than off "the lookup didn't succeed".

Asking a provider for a child it doesn't have also killed the debugger
outright:

  (lldb) frame variable f.whatever
  Failure value returned from cantFail wrapped call
  type has no child named 'whatever'
  UNREACHABLE executed at llvm/include/llvm/Support/Error.h:779!

`createStringErrorV` is a header-inline template, so `StringError::ID` is
instantiated in the caller and liblldb carries its own hidden copy. The
error then fails `ErrorInfoBase::isA` in whichever image consumes it,
`handleErrors` hands it back unhandled, and the `cantFail` inside
`consumeError`/`toString` aborts - whichever one touches it first. Using
the out-of-line `llvm::createStringError` keeps construction inside
libLLVMSupport and fixes it. The same pattern appears elsewhere in LLDB;
only the call site that aborts is changed here.

+++ lldb/include/lldb/Interpreter/Interfaces/ScriptedSyntheticChildrenInterface.h
@@ -27,7 +27,10 @@ public:
     return lldb::ValueObjectSP();
   }
 
-  virtual llvm::Expected<uint32_t> GetIndexOfChildWithName(ConstString name) {
+  // Returns size_t, matching SyntheticChildrenFrontEnd, so the front end can
+  // forward this result without converting between two different
+  // Expected<T> instantiations.
+  virtual llvm::Expected<size_t> GetIndexOfChildWithName(ConstString name) {
     return llvm::createStringErrorV("type has no child named '{0}'", name);
   }
